+++ matpic/str.h
@@ -28,7 +28,7 @@ extern int skipnl(char **src);
 extern char *getid(char **src);
 extern int idlen(char *src);
 extern int cmpid(char *idl, char *idr);
-extern int getnum(char **src, unsigned int *ret);
+extern int getnum(char **src, unsigned long *ret);
 extern int linelen(char *str);
 extern char *mstrldup(char *s, int len);
